backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
httpx==0.27.0
pytest==8.1.1
pytest-django==4.8.0
factory-boy==3.3.0
responses==0.25.0
python-dotenv==1.0.1
django-cors-headers==4.3.1
